update after code review
