+++ tests/integration/targets/helm/tasks/tests_chart.yml
@@ -411,39 +411,6 @@
           - result is failed
           - "'context \"does-not-exist\" does not exist' in result.stderr"
 
-    # Test helm set
-    - name: Install helm using set_values parameters
-      kubernetes.core.helm:
-        binary_path: "{{ helm_binary }}"
-        chart_ref: mariadb
-        chart_repo_url: https://charts.bitnami.com/bitnami
-        release_name: test-mariadb
-        release_namespace: "{{ helm_namespace }}"
-        create_namespace: true
-        set_values:
-          - value: phase=integration
-            value_type: string
-          - value: versioned=false
-
-    - name: Get value set as string
-      kubernetes.core.helm_get:
-        binary_path: "{{ helm_binary }}"
-        release_name: test-mariadb
-        release_namespace: "{{ helm_namespace }}"
-        release_info:
-          - values
-      register: user_values
-
-    - name: Assert that release was created with user-defined variables
-      assert:
-        that:
-          - '"versioned" in release_values'
-          - '"phase" in release_values'
-          - release_values.phase == "integration"
-          - release_values.versioned is false
-      vars:
-        release_values: "{{ user_values['values'] }}"
-
   always:
     - name: Clean up temp dir
       file:
